src/requirements.txt
fastapi
uvicorn
openai
transformers
accelerate
torch
sentence-transformers
faiss-cpu
peft
datasets
python-dotenv
scikit-learn
pydantic
uvicorn 
openai 
transformers 
accelerate 
torch 
sentence-transformers 
faiss-cpu
peft 
datasets 
scikit-learn 
pydantic 
python-dotenv